requirements.txt
# Django
Django>=2.2.10,<2.3.0
django-bootstrap4>=1.1.1,<1.2.0
django-icons>=1.1.1,<1.2.0

# Test and Linter/auto formatting
flake8>=3.7.9,<3.8.0
pytest-django>=3.8.0,<3.9.0
black==19.10b0
coverage>=5.0.3,<5.1.0

# Utilities
factory-boy>=2.11.0,<2.12.0

pytz==2019.3
django-heroku
